Extract `TopUpTModal` to a separate file
